Refactor and fix publications for non-mpp
